Adds packet and trade logging
Currently packet logging doesn't support sending_shard because there's
no easy way to get the port of a sending IP in a socket, and since all
the live shards are now on the same IP, the only way to differentiate
them is through ports.  The easiest way I can think of doing this is to
add another line to packet messages saying the shard, but this has the
disadvantage of adding a line to every packet sent.